+++ CHANGELOG.md
@@ -18,7 +18,8 @@
 16. Added `Fonts`.
 17. Added `pause_upon_start` kwarg for `Clock`.
 18. Added `Clock.set_time()`.
-19. Added `Sound` and `SoundManager`
+19. Added `Sound` and `SoundManager`.
+20. Added `SceneManager.auto_find_scenes()`.
 
 ## version 2.5.0
 ### New Features